+++ src/schemathesis/__init__.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 
 from schemathesis import auths, engine, errors, graphql, hooks, openapi, pytest, python
 from schemathesis.checks import CheckContext, CheckFunction, check
-from schemathesis.core.output import sanitization
 from schemathesis.core.transport import Response
 from schemathesis.core.version import SCHEMATHESIS_VERSION
 from schemathesis.generation import GenerationMode
 from schemathesis.generation.case import Case
-from schemathesis.generation.targets import TargetContext, TargetFunction, target
+from schemathesis.generation.metrics import MetricContext, MetricFunction, metric
 from schemathesis.hooks import HookContext
 from schemathesis.schemas import BaseSchema
+from schemathesis.transport import SerializationContext, serializer
 
 __version__ = SCHEMATHESIS_VERSION
 
@@ -23,8 +23,6 @@ __all__ = [
     "CheckFunction",
     "GenerationMode",
     "Response",
-    "TargetContext",
-    "TargetFunction",
     "HookContext",
     "BaseSchema",
     "__version__",
@@ -38,6 +36,11 @@ __all__ = [
     "openapi",
     "pytest",
     "python",
-    "sanitization",
-    "target",
+    # Targeted Property-based Testing
+    "metric",
+    "MetricContext",
+    "MetricFunction",
+    # Serialization
+    "serializer",
+    "SerializationContext",
 ]